refactor: core dir
