+++ src-frontend/core/components/scratch/index.ts
