+++ src-frontend/core/diff-indicators/diff-indicators.ts
@@ -1,11 +1,11 @@
 /** @file Displays indicators and info on sprites that were changed */
-import type { Project, Sprite } from "./api";
-import { misc, sprites } from "./components";
-import type { DiffModal } from "./modals";
-import { parseScripts } from "./scripts";
-import { SpriteDiff, StageDiff } from "./components";
-import { getBlockly } from "./lib/globals";
-import { userSettings } from "./settings";
+import type { Project, Sprite } from "../api";
+import { misc, sprites } from "../components";
+import type { DiffModal } from "../modals";
+import { parseScripts } from "./script-parser";
+import { SpriteDiff, StageDiff } from "../components";
+import { getBlockly } from "../../lib/globals";
+import { userSettings } from "../settings";
 
 const STAGE = {
   name: "Stage",
