+++ src-frontend/core/diff-indicators/index.ts
@@ -1,0 +1,2 @@
+export { showIndicators } from "./diff-indicators";
+export { type ScriptStatus, parseScripts } from "./script-parser";
