+++ src-frontend/core/modals/commit/styles.css
