+++ src-frontend/core/modals/diff/block-utils.ts
