+++ src-frontend/core/modals/diff/styles.css
