+++ src-frontend/core/modals/thumbnail.svg.ts
