+++ src-frontend/core/modals/welcome/styles.css
