+++ src-frontend/core/settings.ts
