+++ src-frontend/core/utils.ts
